requirements.txt
annotated-types==0.7.0
anyio==4.9.0
authlib==1.6.0
cachecontrol==0.14.3
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
cryptography==45.0.3
deprecated==1.2.18
docstring-parser==0.16
dramatiq==1.18.0
fastapi==0.115.12
firebase-admin==6.9.0
gevent==25.5.1
google-adk==1.4.1
google-api-core==2.25.0
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.97.0
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-logging==3.12.1
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.24.0
google-cloud-speech==2.33.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.2
google-crc32c==1.7.1
google-genai==1.21.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.21
greenlet==3.2.2
grpc-google-iam-v1==0.14.2
grpcio==1.72.1
grpcio-status==1.72.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
importlib-metadata==8.7.0
mcp==1.9.4
msgpack==1.1.1
numpy==2.3.0
opentelemetry-api==1.34.1
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
packaging==25.0
pika==1.3.2
priority==2.0.0
prometheus-client==0.22.1
proto-plus==1.26.1
protobuf==6.31.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic-core==2.33.2
pydantic-settings==2.9.1
pyjwt==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pyyaml==6.0.2
redis==6.2.0
requests==2.32.4
rsa==4.9.1
ruff==0.11.11
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.41
sse-starlette==2.3.6
sseclient-py==1.8.0
starlette==0.46.2
tenacity==8.5.0
typing-extensions==4.13.2
typing-inspection==0.4.1
tzlocal==5.3.1
ulid-py==1.1.0
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.34.3
watchdog==6.0.0
watchdog-gevent==0.2.1
websockets==15.0.1
wrapt==1.17.2
wsproto==1.2.0
zipp==3.23.0
zope-event==5.0
zope-interface==7.2
